requirements.txt
matplotlib~=3.5.1
imageio~=2.16.0
scikit-image~=0.19.1

numpy~=1.21.2
pillow~=8.4.0
torchvision~=0.11.3
tensorboard~=2.8.0
